# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ānvīkṣikī publication platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
        "429":
          description: Rate limited
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current authenticated admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # ─── SITE SETTINGS ────────────────────────────────────────────────────────
  /admin/settings:
    get:
      operationId: getSiteSettings
      tags: [admin]
      summary: Get site settings (admin)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSiteSettings
      tags: [admin]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

  # ─── STATS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminLoginInput:
      type: object
      properties:
        email:    { type: string }
        password: { type: string }
      required: [email, password]
    AdminUser:
      type: object
      properties:
        id:    { type: integer }
        email: { type: string }
        role:  { type: string, enum: [owner, editor, contributor] }
      required: [id, email, role]

    # ─── Settings ────────────────────────────────────────────────────────
    SiteSettings:
      type: object
      properties:
        featuredArticleId: { type: ["integer", "null"] }
        siteTitle:         { type: string }
        siteDescription:   { type: string }
      required: [siteTitle, siteDescription]
    SiteSettingsUpdate:
      type: object
      properties:
        featuredArticleId: { type: ["integer", "null"] }
        siteTitle:         { type: string }
        siteDescription:   { type: string }

    # ─── Stats ───────────────────────────────────────────────────────────
...