# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ānvīkṣikī publication platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  /** @nullable */
  featuredArticleId?: number | null;
  siteTitle: string;
  siteDescription: string;
}